updates to the mlgt register list
from the 2024-11 TMG meeting:

one of the french members has left, their name was removed;

the german poc was updated to be Mr. Chris Schubert from TU Vienna after discussion with the German HoD Markus Seifert;

@HassanAkbar @ronaldtse i noticed a couple of the operating-language-character-encoding:  values were missing

also, in the future, do you  think we should have an english name for each of the registers as register-name: and content-summary: are often in the register language; for english searching it could be helpful to have register-name-en and content-summary-en;  but there is no hurry to consider this; not an enhancement request, only an idea for future work;

if my updates are ok, at your convenience, please push this to the main tc211 geolexica site.  thank you!